+++ src/page/DuesManagement/index.tsx
@@ -1,4 +1,3 @@
-import { useNavigate } from 'react-router-dom';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
@@ -15,16 +14,19 @@ import {
 import { useGetAllDues } from 'query/dues';
 import useBooleanState from 'util/hooks/useBooleanState.ts';
 import { DuesDetail } from 'model/dues/allDues';
-import { LAST_DUES_YEAR, STATUS, STATUS_MAPPING } from 'util/constants/status';
+import { STATUS, STATUS_MAPPING } from 'util/constants/status';
 import { useGetTracks } from 'query/tracks';
 import LoadingSpinner from 'layout/LoadingSpinner';
 import {
-  ArrowBackIosNewOutlined, ArrowDownward, ArrowForwardIosOutlined, ArrowUpward, Sort,
+  ArrowDownward, ArrowUpward, Sort,
 } from '@mui/icons-material';
 import useQueryParam from 'util/hooks/useQueryParam';
 import makeNumberArray from 'util/hooks/makeNumberArray';
 import { useGetMembers } from 'query/members';
 import { useSnackBar } from 'ts/useSnackBar';
+import YearPagination from 'component/YearPagination';
+import useMediaQuery from 'util/hooks/useMediaQuery';
+import PersonalDues from 'page/PersonalDues';
 import * as S from './style';
 
 interface SortAnchorEl {
@@ -33,8 +35,8 @@ interface SortAnchorEl {
 }
 
 function DefaultTable() {
-  const navigate = useNavigate();
-  const page = useQueryParam('page', 'number') as number | null;
+  const param = useQueryParam('page');
+  const page = Number(param);
   const currentYear = new Date().getFullYear();
   const [duesYear, setDuesYear] = useState(page ? currentYear - page + 1 : currentYear);
   const [name, setName] = useState('');
@@ -143,37 +145,16 @@ function DefaultTable() {
   };
 
   useEffect(() => {
-    setDuesYear(page ? currentYear - page + 1 : currentYear);
     if (allDues.dues) {
       setFilteredValue(allDues.dues.filter((row) => members?.content.some((member) => member.memberType === 'REGULAR' && member.id === row.memberId)));
     }
-  }, [currentYear, page, allDues.dues, members?.content]);
+  }, [allDues.dues, members?.content]);
 
-  const goToPrevYear = () => {
-    // 재학생 회비 내역이 2021년부터 시작하므로 2021년 이전으로 이동할 수 없음
-    const prevYear = page ? page + 1 : 2;
-    if (prevYear <= currentYear - 2020) {
-      navigate(`/dues?page=${prevYear}`);
-    }
-  };
-
-  const goToNextYear = () => {
-    if (page && page > 1) {
-      navigate(`/dues?page=${page - 1}`);
-    }
-  };
   return (
     <>
       <div css={S.searchAndPagination}>
         <div css={S.pagination}>
-          {/* 회비 데이터의 마지막 연도는 2021년 입니다. */}
-          <Button onClick={goToPrevYear} disabled={duesYear === LAST_DUES_YEAR}>
-            <ArrowBackIosNewOutlined />
-          </Button>
-          <span css={S.paginationTitle}>{duesYear}</span>
-          <Button onClick={goToNextYear} disabled={duesYear === currentYear}>
-            <ArrowForwardIosOutlined />
-          </Button>
+          <YearPagination duesYear={duesYear} setDuesYear={setDuesYear} />
         </div>
         <div>
           <Input
@@ -360,11 +341,16 @@ function DefaultTable() {
 }
 
 export default function DuesManagement() {
+  const { isMobile } = useMediaQuery();
   return (
     <div css={S.container}>
       <div css={S.mainContent}>
         <Suspense fallback={<LoadingSpinner />}>
-          <DefaultTable />
+          {isMobile ? (
+            <PersonalDues />
+          ) : (
+            <DefaultTable />
+          )}
         </Suspense>
       </div>
     </div>
